Move stories back to the root
