+++ stories/LoginButton.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 
-import LoginButton from '../components/LoginButton';
+import LoginButton from '../src/components/LoginButton';
 
 storiesOf('LoginButton', module).add('default', () => <LoginButton />);
