+++ stories/setup/styles.scss
